export FakePrimitives from fakes
